BatterySimulator/requirements.txt
pybamm[jax]==24.5
fastapi==0.115.2
uvicorn==0.32.0
celery[redis]==5.4.0
pika==1.3.2
bpx==0.4.0
python-dotenv==1.0.0
configparser==7.0.0
pydantic==1.10.10